GW-98: Refactored folder structure
