+++ client/src/pages/games/components/Profile.jsx
