--- client/src/pages/games/dashboard/Dashboard.jsx
@@ -1,11 +1,0 @@
-import Leaderboard from 'pages/games/dashboard/components/Leaderboard';
-import Sidebar from 'pages/games/dashboard/components/Sidebar';
-
-export default function Dashboard() {
-	return (
-		<div className='flex flex-col'>
-			<Sidebar />
-			<Leaderboard />
-		</div>
-	);
-}
